% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algae.R
\docType{package}
\name{algae}
\alias{algae}
\alias{algae-package}
\title{algae: A package for ....}
\description{
The foo package provides three categories of important functions:
AA, BB, CC
}
\section{AA functions}{

The OO functions ...
}

