% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{Al10SABI}
\alias{Al10SABI}
\alias{Am092Bsub}
\title{Al10SABI algorithm}
\usage{
Al10SABI(w857, w644, w458, w529)

Am092Bsub(w678, w667)
}
\arguments{
\item{w857}{numeric. Value at wavelength of 678 nm}

\item{w644}{numeric. Value at wavelength of 667 nm}

\item{w458}{numeric. Value at wavelength of 678 nm}

\item{w529}{numeric. Value at wavelength of 667 nm}

\item{w678}{numeric. Value at wavelength of 678 nm}

\item{w667}{numeric. Value at wavelength of 667 nm}
}
\value{
RasterLayer or numeric

RasterLayer or numeric
}
\description{
Applies the Al10SABI algorithm

Applies the Am092Bsub algorithm
}
\references{
Alawadi, F. Detection of surface algal blooms using the newly developed algorithm surface algal bloom index (SABI). Proc. SPIE 2010, 7825.

Amin, R.; Zhou, J.; Gilerson, A.; Gross, B.; Moshary, F.; Ahmed, S. Novel optical techniques for detecting and classifying toxic dinoflagellate Karenia brevis blooms using satellite imagery. Opt. Express 2009, 17, 9126–9144.
}
