% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{Be16FLHblue}
\alias{Be16FLHblue}
\title{Be16FLHblue algorithm}
\usage{
Be16FLHblue(w529, w644, w458)
}
\arguments{
\item{w529}{numeric. Value at wavelength of 529 nm}

\item{w644}{numeric. Value at wavelength of 644 nm}

\item{w458}{numeric. Value at wavelength of 458 nm}
}
\value{
RasterLayer or numeric
}
\description{
Applies the Be16FLHblue algorithm
}
\references{
Beck, R.A. and 22 others; Comparison of satellite reflectance algorithms for estimating chlorophyll-a in a temperate reservoir using coincident hyperspectral aircraft imagery and dense coincident surface observations, Remote Sens. Environ., 2016, 178, 15-30.
}
\seealso{
Other algorithms: \code{\link{Al10SABI}},
  \code{\link{Am092Bsub}}, \code{\link{Am09KBBI}},
  \code{\link{Be162B643sub629}},
  \code{\link{Be162B700sub601}},
  \code{\link{Be162BsubPhy}},
  \code{\link{Be16FLHBlueRedNIR}},
  \code{\link{Be16FLHGreenRedNIR}},
  \code{\link{Be16FLHVioletRedNIR}},
  \code{\link{Be16FLHviolet}}, \code{\link{Be16MPI}},
  \code{\link{Be16NDPhyI644over615}},
  \code{\link{Be16NDPhyI644over629}},
  \code{\link{Be16NDPhyI}}, \code{\link{Be16NDTIblue}},
  \code{\link{Be16NDTIviolet}},
  \code{\link{Be16Phy2BDA644over629}},
  \code{\link{Da052BDA}}, \code{\link{De933BDA}},
  \code{\link{Gi033BDA}}, \code{\link{Go04MCI}},
  \code{\link{HU103BDA}}, \code{\link{Kn07KIVU}},
  \code{\link{Ku15PhyCI}}, \code{\link{Ku15SLH}},
  \code{\link{MI092BDA}}, \code{\link{MM092BDA}},
  \code{\link{MM12NDCIalt}}, \code{\link{MM12NDCI}},
  \code{\link{MM143BDAopt}}, \code{\link{SI052BDA}},
  \code{\link{SM122BDA}}, \code{\link{SY002BDA}},
  \code{\link{TurbBe16GreenPlusRedBothOverViolet}},
  \code{\link{TurbBe16RedOverViolet}},
  \code{\link{TurbBow06RedOverGreen}},
  \code{\link{TurbChip09NIROverGreen}},
  \code{\link{TurbDox02NIRoverRed}},
  \code{\link{TurbFrohn09GreenPlusRedBothOverBlue}},
  \code{\link{TurbHarr92NIR}},
  \code{\link{TurbLath91RedOverBlue}},
  \code{\link{TurbMoore80Red}}, \code{\link{Wy08CI}},
  \code{\link{Zh10FLH}}
}
