% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{Be16FLHGreenRedNIR_WV2}
\alias{Be16FLHGreenRedNIR_WV2}
\title{Be16FLHGreenRedNIR_WV2 algorithm}
\usage{
Be16FLHGreenRedNIR_WV2(w658, w857, w558)
}
\arguments{
\item{w658}{numeric. Value at wavelength of 658 nm}

\item{w857}{numeric. Value at wavelength of 857 nm}

\item{w558}{numeric. Value at wavelength of 558 nm}
}
\value{
RasterLayer or numeric
}
\description{
Applies the Be16FLHGreenRedNIR_WV2 algorithm
}
\references{
Beck, R.; Xu, M.; Zhan, S.; Liu, H.; Johansen, R.A.; Tong, S.; Yang, B.; Shu, S.; Wu, Q.; Wang, S.; Berling, K.; Murray, A.; Emery, E.; Reif, M.; Harwood, J.; Young, J.; Martin, M.; Stillings, G.; Stumpf, R.; Su, H.; Ye, Z.; Huang, Y.	Comparison of Satellite Reflectance Algorithms for Estimating Phycocyanin Values and Cyanobacterial Total Biovolume in a Temperate Reservoir Using Coincident Hyperspectral Aircraft Imagery and Dense Coincident Surface Observations. Remote Sens. 2017, 9, 538.
}
\seealso{
Other algorithms: 
\code{\link{Al10SABI}()},
\code{\link{Am092Bsub}()},
\code{\link{Am09KBBI}()},
\code{\link{Be162B643sub629}()},
\code{\link{Be162B700sub601}()},
\code{\link{Be162BsubPhy}()},
\code{\link{Be16FLHBlueRedNIR_LS8}()},
\code{\link{Be16FLHBlueRedNIR_MERIS}()},
\code{\link{Be16FLHBlueRedNIR_OLCI}()},
\code{\link{Be16FLHBlueRedNIR_S2}()},
\code{\link{Be16FLHBlueRedNIR_WV2}()},
\code{\link{Be16FLHGreenRedNIR_LS8}()},
\code{\link{Be16FLHGreenRedNIR_MERIS}()},
\code{\link{Be16FLHGreenRedNIR_OLCI}()},
\code{\link{Be16FLHGreenRedNIR_S2}()},
\code{\link{Be16FLHVioletRedNIR_LS8}()},
\code{\link{Be16FLHVioletRedNIR_MERIS}()},
\code{\link{Be16FLHVioletRedNIR_OLCI}()},
\code{\link{Be16FLHVioletRedNIR_S2}()},
\code{\link{Be16FLHVioletRedNIR_WV2}()},
\code{\link{Be16FLHblue_LS8}()},
\code{\link{Be16FLHblue_MERIS}()},
\code{\link{Be16FLHblue_OLCI}()},
\code{\link{Be16FLHblue_S2}()},
\code{\link{Be16FLHblue_WV2}()},
\code{\link{Be16FLHviolet_LS8}()},
\code{\link{Be16FLHviolet_MERIS}()},
\code{\link{Be16FLHviolet_OLCI}()},
\code{\link{Be16FLHviolet_S2}()},
\code{\link{Be16FLHviolet_WV2}()},
\code{\link{Be16NDPhyI644over615}()},
\code{\link{Be16NDPhyI644over629}()},
\code{\link{Be16NDPhyI}()},
\code{\link{Be16NDTIblue}()},
\code{\link{Be16NDTIviolet}()},
\code{\link{Be16Phy2BDA644over629}()},
\code{\link{Da052BDA}()},
\code{\link{De933BDA}()},
\code{\link{Gi033BDA}()},
\code{\link{Go04MCI}()},
\code{\link{HU103BDA}()},
\code{\link{Kn07KIVU}()},
\code{\link{MI092BDA}()},
\code{\link{MM092BDA}()},
\code{\link{MM12NDCIalt}()},
\code{\link{MM12NDCI}()},
\code{\link{MM143BDAopt}()},
\code{\link{SI052BDA}()},
\code{\link{SM122BDA}()},
\code{\link{SY002BDA}()},
\code{\link{TurbBe16GreenPlusRedBothOverViolet}()},
\code{\link{TurbBe16RedOverViolet}()},
\code{\link{TurbBow06RedOverGreen}()},
\code{\link{TurbChip09NIROverGreen}()},
\code{\link{TurbDox02NIRoverRed}()},
\code{\link{TurbFrohn09GreenPlusRedBothOverBlue}()},
\code{\link{TurbHarr92NIR}()},
\code{\link{TurbLath91RedOverBlue}()},
\code{\link{TurbMoore80Red}()},
\code{\link{Wy08CI}()}
}
\concept{algorithms}
