% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{HU103BDA}
\alias{HU103BDA}
\title{HU103BDA algorithm}
\usage{
HU103BDA(w615, w600, w725)
}
\arguments{
\item{w615}{numeric. Value at wavelength of 615 nm}

\item{w600}{numeric. Value at wavelength of 600 nm}

\item{w725}{numeric. Value at wavelength of 725 nm}
}
\value{
RasterLayer or numeric
}
\description{
Applies the HU103BDA algorithm
}
\references{
Hunter, P.D.; Tyler, A.N.; Willby, N.J.; Gilvear, D.J.; The spatial dynamics of vertical migration by Microcystis aeruginosa in a eutrophic shallow lake: A case study using high spatial resolution time-series airborne remote sensing.  Limn. Oceanogr. 2008, 53, 2391-2406.
}
\seealso{
Other algorithms: \code{\link{Al10SABI}},
  \code{\link{Am092Bsub}}, \code{\link{Am09KBBI}},
  \code{\link{Be162B643sub629}},
  \code{\link{Be162B700sub601}},
  \code{\link{Be162BsubPhy}},
  \code{\link{Be16FLHBlueRedNIR}},
  \code{\link{Be16FLHGreenRedNIR}},
  \code{\link{Be16FLHVioletRedNIR}},
  \code{\link{Be16FLHblue}}, \code{\link{Be16FLHviolet}},
  \code{\link{Be16MPI}},
  \code{\link{Be16NDPhyI644over615}},
  \code{\link{Be16NDPhyI644over629}},
  \code{\link{Be16NDPhyI}}, \code{\link{Be16NDTIblue}},
  \code{\link{Be16NDTIviolet}},
  \code{\link{Be16Phy2BDA644over629}},
  \code{\link{Da052BDA}}, \code{\link{De933BDA}},
  \code{\link{Gi033BDA}}, \code{\link{Go04MCI}},
  \code{\link{Kn07KIVU}}, \code{\link{Ku15PhyCI}},
  \code{\link{Ku15SLH}}, \code{\link{MI092BDA}},
  \code{\link{MM092BDA}}, \code{\link{MM12NDCIalt}},
  \code{\link{MM12NDCI}}, \code{\link{MM143BDAopt}},
  \code{\link{SI052BDA}}, \code{\link{SM122BDA}},
  \code{\link{SY002BDA}},
  \code{\link{TurbBe16GreenPlusRedBothOverViolet}},
  \code{\link{TurbBe16RedOverViolet}},
  \code{\link{TurbBow06RedOverGreen}},
  \code{\link{TurbChip09NIROverGreen}},
  \code{\link{TurbDox02NIRoverRed}},
  \code{\link{TurbFrohn09GreenPlusRedBothOverBlue}},
  \code{\link{TurbHarr92NIR}},
  \code{\link{TurbLath91RedOverBlue}},
  \code{\link{TurbMoore80Red}}, \code{\link{Wy08CI}},
  \code{\link{Zh10FLH}}
}
