% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_wq.R
\name{Map_WQ_raster}
\alias{Map_WQ_raster}
\title{Create waterquality Map with sampling points and optional histogram}
\usage{
Map_WQ_raster(
  WQ_raster,
  sample_points,
  map_title,
  raster_style = "quantile",
  histogram = TRUE
)
}
\arguments{
\item{WQ_raster}{raster tiff of waterquality output}

\item{sample_points}{geospatial points object}

\item{map_title}{text for the title of the map}

\item{raster_style}{data visualization style (default is quantile)}

\item{histogram}{boolean toggle to add or remove histogram (default is TRUE)}
}
\value{
A data visualization of the results
}
\description{
This function wraps the tmap package to help users generate fast and simple
data visualization of their WQ_calc raster output along with optional
geospatial objects and histogram
}
\seealso{
Other Map_WQ models: 
\code{\link{Map_WQ_basemap}()}
}
\concept{Map_WQ models}
