% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_correlations.R
\name{plot_hirsh_vs_theobald}
\alias{plot_hirsh_vs_theobald}
\title{Plot to compare Hirsh and Theobald datasets}
\usage{
plot_hirsh_vs_theobald(
  master_data = wq_prepare_data("master_data"),
  outdir = "figs"
)
}
\arguments{
\item{master_data}{Master dataset.}

\item{outdir}{Output directory (where figures are saved).}
}
\description{
Plot to compare Hirsh and Theobald datasets
}
