% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run the Shiny Application}
\usage{
run_app(..., options.shiny = list())
}
\arguments{
\item{...}{arguments to pass to golem_opts.
See \code{?golem::get_golem_options} for more details.}

\item{options.shiny}{A list of options to pass to \code{shiny::shinyApp()}.}
}
\description{
This function launches the Water Quality Logistic Regression Explorer Shiny application.
The app allows users to interactively build and explore logistic regression models
using any numeric columns from the master dataset.
}
