% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{SY002BDA}
\alias{SY002BDA}
\title{SY002BDA algorithm}
\usage{
SY002BDA(w650, w625)
}
\arguments{
\item{w650}{numeric. Value at wavelength of 650 nm}

\item{w625}{numeric. Value at wavelength of 625 nm}
}
\value{
RasterLayer or numeric
}
\description{
Applies the SY002BDA algorithm
}
\references{
Schalles, J.; Yacobi, Y. Remote detection and seasonal patterns of phycocyanin, carotenoid and chlorophyll-a pigments in eutrophic waters. Archiv fur Hydrobiologie, Special Issues Advances in Limnology, 2000, 55,153–168.
}
\seealso{
Other algorithms: \code{\link{Al10SABI}},
  \code{\link{Am092Bsub}}, \code{\link{Am09KBBI}},
  \code{\link{Be162B643sub629}},
  \code{\link{Be162B700sub601}},
  \code{\link{Be162BsubPhy}},
  \code{\link{Be16FLHBlueRedNIR}},
  \code{\link{Be16FLHGreenRedNIR}},
  \code{\link{Be16FLHVioletRedNIR}},
  \code{\link{Be16FLHblue}}, \code{\link{Be16FLHviolet}},
  \code{\link{Be16MPI}},
  \code{\link{Be16NDPhyI644over615}},
  \code{\link{Be16NDPhyI644over629}},
  \code{\link{Be16NDPhyI}}, \code{\link{Be16NDTIblue}},
  \code{\link{Be16NDTIviolet}},
  \code{\link{Be16Phy2BDA644over629}},
  \code{\link{Da052BDA}}, \code{\link{De933BDA}},
  \code{\link{Gi033BDA}}, \code{\link{Go04MCI}},
  \code{\link{HU103BDA}}, \code{\link{Kn07KIVU}},
  \code{\link{Ku15PhyCI}}, \code{\link{Ku15SLH}},
  \code{\link{MI092BDA}}, \code{\link{MM092BDA}},
  \code{\link{MM12NDCIalt}}, \code{\link{MM12NDCI}},
  \code{\link{MM143BDAopt14}}, \code{\link{MM143BDAopt}},
  \code{\link{SI052BDA}}, \code{\link{SM122BDA}},
  \code{\link{TurbBe16GreenPlusRedBothOverViolet}},
  \code{\link{TurbBe16RedOverViolet}},
  \code{\link{TurbBow06RedOverGreen}},
  \code{\link{TurbChip09NIROverGreen}},
  \code{\link{TurbDox02NIRoverRed}},
  \code{\link{TurbFrohn09GreenPlusRedBothOverBlue}},
  \code{\link{TurbHarr92NIR}},
  \code{\link{TurbLath91RedOverBlue}},
  \code{\link{TurbMoore80Red}}, \code{\link{Wy08CI}},
  \code{\link{Zh10FLH}}
}
