% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{TurbHarr92NIR}
\alias{TurbHarr92NIR}
\title{TurbHarr92NIR     algorithm}
\usage{
TurbHarr92NIR(w857)
}
\arguments{
\item{w857}{numeric. Value at wavelength of 857 nm}
}
\value{
RasterLayer or numeric
}
\description{
Applies the TurbHarr92NIR     algorithm
}
\references{
Schiebe F.R., Harrington J.A., Ritchie J.C. Remote-Sensing of Suspended Sediments—the Lake Chicot, Arkansas Project. Int. J. Remote Sens. 1992;13:1487–1509.
}
\seealso{
Other algorithms: 
\code{\link{Al10SABI}()},
\code{\link{Am092Bsub}()},
\code{\link{Am09KBBI}()},
\code{\link{Be162B643sub629}()},
\code{\link{Be162B700sub601}()},
\code{\link{Be162BsubPhy}()},
\code{\link{Be16FLHBlueRedNIR}()},
\code{\link{Be16FLHGreenRedNIR}()},
\code{\link{Be16FLHVioletRedNIR}()},
\code{\link{Be16FLHblue}()},
\code{\link{Be16FLHviolet}()},
\code{\link{Be16MPI}()},
\code{\link{Be16NDPhyI644over615}()},
\code{\link{Be16NDPhyI644over629}()},
\code{\link{Be16NDPhyI}()},
\code{\link{Be16NDTIblue}()},
\code{\link{Be16NDTIviolet}()},
\code{\link{Be16Phy2BDA644over629}()},
\code{\link{Da052BDA}()},
\code{\link{De933BDA}()},
\code{\link{Gi033BDA}()},
\code{\link{Go04MCI}()},
\code{\link{HU103BDA}()},
\code{\link{Kn07KIVU}()},
\code{\link{Ku15PhyCI}()},
\code{\link{Ku15SLH}()},
\code{\link{MI092BDA}()},
\code{\link{MM092BDA}()},
\code{\link{MM12NDCIalt}()},
\code{\link{MM12NDCI}()},
\code{\link{MM143BDAopt}()},
\code{\link{SI052BDA}()},
\code{\link{SM122BDA}()},
\code{\link{SY002BDA}()},
\code{\link{TurbBe16GreenPlusRedBothOverViolet}()},
\code{\link{TurbBe16RedOverViolet}()},
\code{\link{TurbBow06RedOverGreen}()},
\code{\link{TurbChip09NIROverGreen}()},
\code{\link{TurbDox02NIRoverRed}()},
\code{\link{TurbFrohn09GreenPlusRedBothOverBlue}()},
\code{\link{TurbLath91RedOverBlue}()},
\code{\link{TurbMoore80Red}()},
\code{\link{Wy08CI}()},
\code{\link{Zh10FLH}()}
}
\concept{algorithms}
