% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality-package.R
\docType{package}
\name{waterquality-package}
\alias{waterquality}
\alias{waterquality-package}
\title{waterquality: Satellite Derived Water Quality Detection Algorithms}
\description{
The goal of 'waterquality' is to convert satellite-reflectance imagery
to a host of pre-defined water quality algorithms designed for the detection of
cholorophyll-a, blue-green algae (Phycocyanin), and turbidity.
This package is able to process the following sensor configurations:
WorldView-2, Sentinel-2, Landsat-8, MODIS, and MERIS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RAJohansen/waterquality}
  \item Report bugs at \url{https://github.com/RAJohansen/waterquality/issues}
}

}
\author{
\strong{Maintainer}: Richard Johansen \email{richard.johansen@uc.edu} (0000-0003-4287-9677)

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)
}

Other contributors:
\itemize{
  \item U.S. Army Corps of Engineers [funder]
}

}
\keyword{internal}
