% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality.R
\docType{package}
\name{waterquality}
\alias{waterquality}
\alias{waterquality-package}
\title{waterquality: A package for ....}
\description{
The waterquality package provides three categories of important functions:
AA, BB, CC
}
\section{AA functions}{

The OO functions ...
}

