% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq_calc.R
\name{wq_calc}
\alias{wq_calc}
\title{Water quality calculation}
\usage{
wq_calc(raster_stack, alg = "all", sat = "sentinel2")
}
\arguments{
\item{raster_stack}{RasterStack containing a satellite data}

\item{alg}{Name of the algorithm or "all"}

\item{sat}{Name of the satellite}
}
\value{
RasterLayer
}
\description{
Calculates a set of water quality indices
}
\examples{
\dontrun{(
library(raster)
s2 = stack(system.file("raster/S2_Taylorsville.tif", package = "algae"))
s2_wq = wq_calc(s2)
)}
}
