% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{run_analysis}
\alias{run_analysis}
\title{Run analysis}
\usage{
run_analysis(prepare_data = FALSE)
}
\arguments{
\item{prepare_data}{A logical. Should the steps to prepare data be run?}
}
\description{
Run analysis
}
