% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse_components.R
\name{run_analysis_components}
\alias{run_analysis_components}
\title{Run analysis}
\usage{
run_analysis_components(prepare_data = FALSE)
}
\arguments{
\item{prepare_data}{A logical. Should the steps to prepare data be run?}
}
\description{
Run analysis
}
