% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality-package.R
\docType{package}
\name{waterquality-package}
\alias{waterquality}
\alias{waterquality-package}
\title{waterquality: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sia Chawla \email{scha0351@student.monash.edu}

}
\keyword{internal}
