% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality-package.R
\docType{package}
\name{waterquality-package}
\alias{waterquality}
\alias{waterquality-package}
\title{waterquality: Interactive Yarra River Water Quality Explorer}
\description{
R package containing a cleaned Yarra River water-quality dataset and a Shiny app to explore parameters over time (single-plot multi-parameter overlay), with summaries and documentation. Includes a known 1995–2019 data gap note and shading.
}
\author{
\strong{Maintainer}: Sia Chawla \email{scha0351@student.monash.edu}

}
\keyword{internal}
