% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waterquality-package.R
\docType{package}
\name{waterquality-package}
\alias{waterquality}
\alias{waterquality-package}
\title{waterquality: Satellite Derived Water Quality Detection Algorithms}
\description{
The main purpose of waterquality is to quickly and easily convert
    satellite-based reflectance imagery into one or many well-known water quality
    algorithms designed for the detection of harmful algal blooms or the following
    pigment proxies: chlorophyll-a, blue-green algae (Phycocyanin), and Turbidity.
    (Johansen et al. (2018) <doi:10.1016/j.hal.2018.05.001>).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RAJohansen/waterquality}
  \item Report bugs at \url{https://github.com/RAJohansen/waterquality/issues}
}

}
\author{
\strong{Maintainer}: Richard Johansen \email{richard.a.johansen@erdc.dren.mil} (\href{https://orcid.org/0000-0003-4287-9677}{ORCID})

Authors:
\itemize{
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (\href{https://orcid.org/0000-0002-1057-3721}{ORCID})
  \item Molly Reif \email{Molly.K.Reif@usace.army.mil} (\href{https://orcid.org/0000-0003-4207-1459}{ORCID})
  \item Erich Emery \email{Erich.B.Emery@usace.army.mil}
}

Other contributors:
\itemize{
  \item U.S. Army Corps of Engineers [funder]
}

}
\keyword{internal}
