% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq_algorithms.R
\docType{data}
\name{wq_algorithms}
\alias{wq_algorithms}
\title{wq_algorithms database}
\format{A tibble with 23 rows and 3 variables:
\itemize{
\item name: algorithm name
\item funs: algorithm function
\item sentinel2: list of the bands used from the Sentinel 2 satellite
}}
\usage{
wq_algorithms
}
\description{
A dataset containing the information about the water quality algorithms
}
\keyword{datasets}
