% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq_algorithms.R
\docType{data}
\name{wq_algorithms}
\alias{wq_algorithms}
\title{wq_algorithms database}
\format{
A tibble with 91 rows and 4 variables:
\itemize{
\item name: algorithm name
\item funs: algorithm function
\item satellite: satellite/instrument name ("worldview2", "sentinel2", "landsat8", "modis", or "meris")
\item bands: list of the bands used from the given satellite/instrument
}
}
\usage{
wq_algorithms
}
\description{
A dataset containing the information about the water quality algorithms
}
\keyword{datasets}
