% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq_calc.R
\name{wq_calc}
\alias{wq_calc}
\title{Water quality calculation}
\usage{
wq_calc(raster_stack, alg = "all", sat = "sentinel2")
}
\arguments{
\item{raster_stack}{RasterStack containing a satellite data}

\item{alg}{Name of the algorithm or "all"}

\item{sat}{Name of the satellite or instrument ("worldview2", "sentinel2", "landsat8", "modis", or "meris")}
}
\value{
RasterLayer
}
\description{
Calculates a set of water quality indices
}
\examples{
\dontrun{(
library(raster)
s2 = stack(system.file("raster/S2_Taylorsville.tif", package = "algae"))
s2_wq = wq_calc(s2, alg = "all", sat = "sentinel2")
l8 = stack(system.file("raster/L8_Taylorsville.tif", package = "algae"))
l8_wq = wq_calc(s2, alg = "all", sat = "landsat8")

# errors
s2_wq1 = wq_calc(s2[[1]], alg = "all", sat = "sentinel2")
s2_wq2 = wq_calc(s2, alg = "all", sat = "my_satellite")
s2b = s2[[1:4]]
s2_wq3 = wq_calc(s2b, alg = "all", sat = "my_satellite")
)}
}
