% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{wq_prepare_data}
\alias{wq_prepare_data}
\title{Read data}
\usage{
wq_prepare_data(
  type = c("water_chemistry", "hydrobasins_sites", "hydrobasins_lvl07",
    "hydrobasins_lvl12")
)
}
\arguments{
\item{type}{data type.}
}
\description{
Read data
}
