% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithms.R
\name{Wy08CI}
\alias{Wy08CI}
\title{Wy08CI  algorithm}
\usage{
Wy08CI(w686, w672, w715)
}
\arguments{
\item{w686}{numeric. Value at wavelength of 686 nm}

\item{w672}{numeric. Value at wavelength of 672 nm}

\item{w715}{numeric. Value at wavelength of 715 nm}
}
\value{
RasterLayer or numeric
}
\description{
Applies the Wy08CI  algorithm
}
\references{
Wynne, T. T., Stumpf, R. P., Tomlinson, M. C., Warner, R. A., Tester, P. A., Dyble, J.; Relating spectral shape to cyanobacterial blooms in the Laurentian Great Lakes. Int. J. Remote Sens., 2008, 29, 3665–3672.
}
\seealso{
Other algorithms: \code{\link{Al10SABI}},
  \code{\link{Am092Bsub}}, \code{\link{Am09KBBI}},
  \code{\link{Be162B643sub629}},
  \code{\link{Be162B700sub601}},
  \code{\link{Be162BsubPhy}},
  \code{\link{Be16FLHBlueRedNIR}},
  \code{\link{Be16FLHGreenRedNIR}},
  \code{\link{Be16FLHVioletRedNIR}},
  \code{\link{Be16FLHblue}}, \code{\link{Be16FLHviolet}},
  \code{\link{Be16MPI}},
  \code{\link{Be16NDPhyI644over615}},
  \code{\link{Be16NDPhyI644over629}},
  \code{\link{Be16NDPhyI}}, \code{\link{Be16NDTIblue}},
  \code{\link{Be16NDTIviolet}},
  \code{\link{Be16Phy2BDA644over629}},
  \code{\link{Da052BDA}}, \code{\link{De933BDA}},
  \code{\link{Gi033BDA}}, \code{\link{Go04MCI}},
  \code{\link{HU103BDA}}, \code{\link{Kn07KIVU}},
  \code{\link{Ku15PhyCI}}, \code{\link{Ku15SLH}},
  \code{\link{MI092BDA}}, \code{\link{MM092BDA}},
  \code{\link{MM12NDCIalt}}, \code{\link{MM12NDCI}},
  \code{\link{MM143BDAopt}}, \code{\link{SI052BDA}},
  \code{\link{SM122BDA}}, \code{\link{SY002BDA}},
  \code{\link{TurbBe16GreenPlusRedBothOverViolet}},
  \code{\link{TurbBe16RedOverViolet}},
  \code{\link{TurbBow06RedOverGreen}},
  \code{\link{TurbChip09NIROverGreen}},
  \code{\link{TurbDox02NIRoverRed}},
  \code{\link{TurbFrohn09GreenPlusRedBothOverBlue}},
  \code{\link{TurbHarr92NIR}},
  \code{\link{TurbLath91RedOverBlue}},
  \code{\link{TurbMoore80Red}}, \code{\link{Zh10FLH}}
}
